geotab-mcp-server/requirements.txt
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
mygeotab==0.8.8
